corrected errors in selection_sort v1

+++ 2-selection_sort.c
@@ -11,7 +11,7 @@
 
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min;
+	size_t i, j, min;
 
 	for (i = 0; i < size - 1; i++)
 	{
